requirements.txt
Python 3.8+
PyTorch
matplotlib
numpy
scikit-learn
pandas
